Move samples of int16_t format to sample_t, that is of type double
This prepares the correction of all levels

+++ src/common/emphasis.h
@@ -16,6 +16,6 @@ typedef struct emphasis {
 #define CUT_OFF_EMPHASIS_DEFAULT	300.0
 
 int init_emphasis(emphasis_t *state, int samplerate, double cut_off);
-void pre_emphasis(emphasis_t *state, int16_t *samples, int num);
-void de_emphasis(emphasis_t *state, int16_t *samples, int num);
+void pre_emphasis(emphasis_t *state, double *samples, int num);
+void de_emphasis(emphasis_t *state, double *samples, int num);
